requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-cors>=4.0.0
pydantic>=2.0.0
redis>=4.6.0
pymongo>=4.5.0
sqlalchemy>=2.0.0
openai>=1.0.0
anthropic>=0.3.0
pinecone-client>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
flask-jwt-extended>=4.7.1
pydantic-settings>=2.8.1
pymysql>=1.1.0
tiktoken>=0.9.0
cryptography>=44.0.2
PyJWT==2.7.0
volcengine-python-sdk>=1.0.130
httpx>=0.28.1
bs4>=0.0.2
flasgger>=0.9.7.1
feedparser>=6.0.11
gunicorn>=21.2.0
pytz>=2023.3
firebase-admin>=6.7.0